Revert "Binance feature"
