--- help.txt
@@ -1,19 +1,0 @@
-Основная механика бота заключается в отслеживании конкретные торговых инструментов, которые каждый пользователь может добавить вручную. Для каждого отслеживаемого инструмента каждые 3 минуты бот проверяет актуальную цену на бирже, в случае если цена превышает нижний или верхний порог (которые также устанавливаются пользователем) - пользователь получает соответствующее оповещение, после чего пользователь может либо прекратить отслеживание инструмента, либо изменить достигнутую цену. Бот для своей работы использует библиотеку finance.yahoo.com, что крайне важно учитывать при работе с ним, т.к. многие (почти все) идентификаторы инструментов, торгуемых на MOEX, дополняются символами, обозначающих биржу, например идентификатор яндекса на MOEX - YNDX, на Yahoo - YNDX.ME. Стоит учитывать также тот момент, что из-за довольно большого времени, требуемого для обращения к серверам Yahoo из России, некоторые сообщения могут выводиться не слишком быстро (1-3 секунды, если на конкретном этапе подразумевается обращение к Yahoo).
-
-Подробнее о доступных командах:
-
-1. "Добавить" (/add) - запускает процесс добавление нового инструмента, при этом пользователь должен указать корректный идентификатор инструмента, а также 3 стоимости - начальная (т.е. та, с которой пользователь хочет начать отсчет, в дальнейшем она нужна для отображения глобального изменения стоимости), целевая/максимальная и минимальная. При этом не обязательно указывать все 3 типа стоимости, а только те, которые желает отслеживать пользователей (можно не указывать ни одну из них).
-
-2. "Мои инструменты" (/show) - поочередно выводит отслеживаемые пользователем инструменты, при этом для каждого инструмента помимо указанных пользователем цен также выводится процентное изменение цены за сегодня и за всё время отслеживания (при условии, что пользователь указал стартовую стоимость для инструмента).
-
-3. "Изменить/удалить" (/edit) - позволяет удалить инструмент или изменить любую из 2-х стоимостей конкретного уже отслеживаемого инструмента пользователя, либо прекратить отслеживание любой из ранее указанных стоимостей.
-
-upd:
-
-Добавлен раздел взаимодействия с  криптовалютной биржей Binance. Для подключения в файл настроек setting.py  необходимо добавить API-ключи из личного кабинета Binance.
-
-1. "Текущий курс" предоставляет цену заданного тикера в формате "BTC USDT".
-
-2. "Узнать баланс"  предоставляет текущий баланс пользователя на бирже Binance.
-
-3. "Создать ордер" позволяет создать  Limit order или Market order.
