+++ keyboards.py
@@ -1,8 +1,4 @@
-from telegram import ReplyKeyboardMarkup
-
-KEYBOARD_PERCENT_POOL = ['25%', '50%', '75%', '100%']
-ORDERS_TYPE = ['Limit order', 'Market order']
-ORDERS_SIDE = ['buy', 'sell']
+﻿from telegram import ReplyKeyboardMarkup
 
 
 def main_menu_keyboard():
@@ -59,33 +55,3 @@ def another_pair_keyboard():
     return ReplyKeyboardMarkup([
         ['Другая пара', 'Отмена']
     ], one_time_keyboard=True, row_width=1, resize_keyboard=True)
-
-
-def order_type_keyboard():
-    return ReplyKeyboardMarkup([
-        ['Limit order', 'Market order'],
-        ['Отмена']
-    ], one_time_keyboard=True, row_width=1, resize_keyboard=True)
-
-
-def buy_sell_keyboard(balance_ticker_1, balance_ticker_2):
-    return ReplyKeyboardMarkup([
-        [f'Доступно {balance_ticker_1}', f'Доступно {balance_ticker_2}'],
-        ['buy', 'sell'],
-        ['Отмена'],
-    ], one_time_keyboard=False, row_width=1, resize_keyboard=True)
-
-
-def quantity_keyboard():
-    return ReplyKeyboardMarkup([
-        ['25%', '50%', '75%', '100%'],
-        ['Отмена']
-    ], one_time_keyboard=True, row_width=1, resize_keyboard=True)
-
-
-def aply_order_keyboard(order_side):
-    return ReplyKeyboardMarkup([
-        [f'{order_side}'],
-        ['Отмена']
-    ], one_time_keyboard=True, row_width=1, resize_keyboard=True)
-
