+++ main.py
@@ -2,12 +2,9 @@ import logging
 import settings
 import telegram
 
-from handlers_binance_calls import get_balance, get_price
-from handlers_binance_calls import get_step_1, get_step_2
 
-from handlers_binance_set_order import set_order, choosing_order_type
-from handlers_binance_set_order import choosing_pair, choosing_order_side
-from handlers_binance_set_order import prepearing_order, making_order, checking_price
+from handlers_binance_calls import binance_comands, get_balance, get_price
+from handlers_binance_calls import get_step_1, get_step_2
 
 from handlers_asset_add import add_start, add_step_1
 from handlers_asset_add import add_step_2, add_step_3, add_step_4
@@ -18,7 +15,7 @@ from handlers_asset_edit_del import (
 from handlers_asset_view import asset_view
 from handlers_utils import (
     greet_user, unknown_text, operation_cancel, show_help,
-    shares_comands, binance_comands
+    shares_comands
 )
 
 from settings import TELEGRAM_API_KEY
@@ -133,40 +130,6 @@ def main():
                     Filters.text & (~Filters.regex('(Отмена)')), get_step_2
                 )
             ]
-
-    orders = ConversationHandler(
-        entry_points=[MessageHandler(Filters.regex('Создать ордер'), set_order)],
-        states={
-            "set_step_1": [
-                MessageHandler(
-                    Filters.text & (~Filters.regex('(Отмена)')), choosing_pair
-                )
-            ],
-            "set_step_2": [
-                MessageHandler(
-                    Filters.text & (~Filters.regex('(Отмена)')), choosing_order_type
-                )
-            ],
-            "set_step_3": [
-                MessageHandler(
-                    Filters.text & (~Filters.regex('(Отмена)')), choosing_order_side
-                )
-            ],
-            "set_step_4": [
-                MessageHandler(
-                    Filters.text & (~Filters.regex('(Отмена)')), checking_price
-                )
-            ],
-            "set_step_5": [
-                MessageHandler(
-                    Filters.text & (~Filters.regex('(Отмена)')), prepearing_order
-                )
-            ],
-            "set_step_6": [
-                MessageHandler(
-                    Filters.text & (~Filters.regex('(Отмена)')), making_order
-                )
-            ],
 
         },
         fallbacks=[MessageHandler(Filters.regex('(Отмена)'), operation_cancel)]
